Develop

See merge request sulinos/devel/inary!27

+++ inary/archive.py
@@ -194,7 +194,10 @@ class ArchiveBinary(ArchiveBase):
         # directory and leave the dirty job to actions.py ;)
         target_file = os.path.join(target_dir,
                                    os.path.basename(self.file_path))
-        shutil.copyfile(self.file_path, target_file)
+        if os.path.isfile(self.file_path):
+            shutil.copyfile(self.file_path, target_file)
+        elif os.path.isdir(self.file_path):
+            shutil.copytree(self.file_path, target_file)
 
 
 class ArchiveBzip2(ArchiveBase):
@@ -896,6 +899,8 @@ class SourceArchive:
             ctx.config.archives_dir(), self.url.filename())
         self.archive = archive
         self.progress = None
+        self.isgit=(self.url.get_uri().startswith("git://") or self.url.get_uri().endswith(".git"))
+        self.branch="master" # TODO need support branch from pspec
 
     def fetch(self, interactive=True):
         if not self.is_cached(interactive):
@@ -910,6 +915,9 @@ class SourceArchive:
                     self.fetch_from_mirror()
                 elif self.url.get_uri().startswith("file://") or self.url.get_uri().startswith("/"):
                     self.fetch_from_locale()
+                elif self.isgit:
+                    self.branch=self.archive.sha1sum
+                    inary.fetcher.fetch_git(self.url,ctx.config.archives_dir()+"/"+self.url.filename(),self.branch)
                 else:
                     inary.fetcher.fetch_url(
                         self.url, ctx.config.archives_dir(), self.progress)
@@ -944,6 +952,7 @@ class SourceArchive:
             self.progress)
 
     def is_cached(self, interactive=True):
+    
         if not os.access(self.archiveFile, os.R_OK):
             return False
 
@@ -958,6 +967,7 @@ class SourceArchive:
         return False
 
     def unpack(self, target_dir, clean_dir=True):
+
 
         # check archive file's integrity
         if not util.check_file_hash(self.archiveFile, self.archive.sha1sum):
